+++ inst/sql/result.sql
@@ -1,0 +1,1 @@
+SELECT state, result_rtype FROM canard_absurd.tasks WHERE id = ?id;